# docs/swagger.yaml
definitions:
...
  models.ProductOptions:
    properties:
      id:
        type: integer
      is_required:
        type: boolean
      is_unique:
        description: define if the option is single or multiple choise
        type: string
      name:
        type: string
      product_id:
        type: integer
    type: object
...
  models.Roles:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      permissions:
        type: string
      role_name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Super Ordening System
  title: WacDo
  version: "1.0"
paths:
...
  /options:
    get:
      description: Retrieve a list of all product options
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProductOptions'
            type: array
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all product options
      tags:
      - Product Options
    post:
      consumes:
      - application/json
      description: Create a new option for a product (e.g. Size, Toppings)
      parameters:
      - description: Option details
        in: body
        name: option
        required: true
        schema:
          $ref: '#/definitions/models.ProductOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductOptions'
        "400":
          description: Invalid data or product not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new product option
      tags:
      - Product Options
  /options/{id}:
    delete:
      description: Delete a product option by ID (also deletes its values)
      parameters:
      - description: Option ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Option deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Option not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a product option
      tags:
      - Product Options
    get:
      description: Retrieve a single product option by its ID
      parameters:
      - description: Option ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductOptions'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Option not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a product option by ID
      tags:
      - Product Options
    put:
      consumes:
      - application/json
      description: Update an existing product option by ID
      parameters:
      - description: Option ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated option details
        in: body
        name: option
        required: true
        schema:
          $ref: '#/definitions/models.ProductOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProductOptions'
        "400":
          description: Invalid ID or data
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Option not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Option name already exists for this product
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a product option
      tags:
      - Product Options
...
  /options/product/{product_id}:
    get:
      description: Retrieve all options belonging to a specific product
      parameters:
      - description: Product ID
        in: path
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ProductOptions'
            type: array
        "400":
          description: Invalid product ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get options by product
      tags:
      - Product Options
...
  /roles:
    get:
      description: Retrieve a list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Roles'
            type: array
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role with the provided details
      parameters:
      - description: Role details
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Roles'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Roles'
        "400":
          description: Invalid data or role already exists
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new role
      tags:
      - Roles
  /roles/{id}:
    delete:
      description: Delete a role by ID (only if not in use by any users)
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Role still in use
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a role
      tags:
      - Roles
    get:
      description: Retrieve a single role by its ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Roles'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a role by ID
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"